+++ backend/development/src/main/kotlin/io/tolgee/util/InMemoryFileStorage.kt
@@ -34,6 +34,7 @@ class InMemoryFileStorage : FileStorage {
   override fun storeFile(
     storageFilePath: String,
     bytes: ByteArray,
+    contentType: String?,
   ) {
     files[storageFilePath] = bytes
   }